+++ src/data/loom-file.ts
@@ -1,7 +1,7 @@
 import { App, Notice, normalizePath } from "obsidian";
 import { createFile, createFolder } from "./file-operations";
 import { createLoomState } from "../shared/loom-state/loom-state-factory";
-import { serializeLoomState } from "./serialize-loom-state";
+import { serializeLoomState } from "./serialize";
 import { FILE_EXTENSION, DEFAULT_LOOM_NAME } from "./constants";
 
 const getFileName = (): string => {
