+++ src/react/export-app/index.tsx
@@ -14,7 +14,7 @@ import {
 } from "../../shared/export/download-utils";
 import { exportToCSV } from "src/shared/export/export-to-csv";
 import { css } from "@emotion/react";
-import { useAppSelector } from "src/redux/global/hooks";
+import { useAppSelector } from "src/redux/hooks";
 
 interface Props {
 	app: App;
