+++ src/react/loom-app/app/use-cell.ts
@@ -1,7 +1,7 @@
-import { useLogger } from "../../../shared/logger";
+import { useLogger } from "src/shared/logger";
 import { useLoomState } from "../loom-state-provider";
-import CellBodyUpdateCommand from "../../../shared/loom-state/commands/cell-body-update-command";
-import CellHeaderUpdateCommand from "../../../shared/loom-state/commands/cell-header-update-command";
+import CellBodyUpdateCommand from "src/shared/loom-state/commands/cell-body-update-command";
+import CellHeaderUpdateCommand from "src/shared/loom-state/commands/cell-header-update-command";
 import React from "react";
 
 export const useCell = () => {
