+++ src/react/loom-app/app/use-row.ts
@@ -1,7 +1,7 @@
-import { useLogger } from "../../../shared/logger";
+import { useLogger } from "src/shared/logger";
 import { useLoomState } from "../loom-state-provider";
-import RowAddCommand from "../../../shared/loom-state/commands/row-add-command";
-import RowDeleteCommand from "../../../shared/loom-state/commands/row-delete-command";
+import RowAddCommand from "src/shared/loom-state/commands/row-add-command";
+import RowDeleteCommand from "src/shared/loom-state/commands/row-delete-command";
 import React from "react";
 
 export const useRow = () => {
