+++ src/react/loom-app/footer-cell-container/calculation.ts
@@ -9,7 +9,8 @@ import {
 	Tag,
 } from "src/shared/loom-state/types";
 import { hashString, round2Digits } from "./utils";
-import { RowNotFoundError, TagNotFoundError } from "src/shared/loom-error";
+import RowNotFoundError from "src/shared/error/row-not-found-error";
+import TagNotFoundError from "src/shared/error/tag-not-found-error";
 import { unixTimeToDateTimeString } from "src/shared/date/date-conversion";
 
 export const getCalculationContent = (
