+++ src/react/loom-app/tag-color-menu/index.tsx
@@ -4,7 +4,7 @@ import Menu from "src/react/shared/menu";
 import ColorItem from "./components/color-item";
 import Text from "src/react/shared/text";
 
-import { useAppSelector } from "src/redux/global/hooks";
+import { useAppSelector } from "src/redux/hooks";
 
 import Divider from "../../shared/divider";
 import Stack from "../../shared/stack";
