+++ src/react/shared/tag/index.tsx
@@ -7,7 +7,7 @@ import Padding from "../padding";
 
 import { Color } from "src/shared/loom-state/types";
 import { findColorClassName } from "src/shared/color";
-import { useAppSelector } from "src/redux/global/hooks";
+import { useAppSelector } from "src/redux/hooks";
 
 interface Props {
 	id?: string;
