+++ src/redux/global-slice.ts
