+++ src/shared/error/column-not-found-error.ts
@@ -1,0 +1,6 @@
+export default class ColumNotFoundError extends Error {
+	constructor(id?: string) {
+		super(`Column ${id} was not found`);
+		this.name = "ColumNotFoundError";
+	}
+}
