+++ src/shared/error/row-not-found-error.ts
@@ -1,0 +1,6 @@
+export default class RowNotFoundError extends Error {
+	constructor(id?: string) {
+		super(`Row ${id} was not found`);
+		this.name = "RowNotFoundError";
+	}
+}
