+++ src/shared/error/tag-not-found-error.ts
@@ -1,0 +1,6 @@
+export default class TagNotFoundError extends Error {
+	constructor(id: string) {
+		super(`Tag ${id} was not found`);
+		this.name = "TagNotFoundError";
+	}
+}
