+++ src/shared/loom-state/commands/cell-footer-update-command.ts
@@ -1,4 +1,4 @@
-import { CellNotFoundError } from "../../loom-error";
+import CellNotFoundError from "src/shared/error/cell-not-found-error";
 import LoomStateCommand from "../loom-state-command";
 import { FooterCell, LoomState } from "../types";
