+++ src/shared/loom-state/commands/cell-header-update-command.ts
@@ -1,4 +1,4 @@
-import { CellNotFoundError } from "../../loom-error";
+import CellNotFoundError from "src/shared/error/cell-not-found-error";
 import LoomStateCommand from "../loom-state-command";
 import { HeaderCell, LoomState } from "../types";
