+++ src/shared/loom-state/commands/tag-update-command.ts
@@ -1,4 +1,4 @@
-import { TagNotFoundError } from "../../loom-error";
+import TagNotFoundError from "src/shared/error/tag-not-found-error";
 import LoomStateCommand from "../loom-state-command";
 import { LoomState, Tag } from "../types";
